Initial checkin.
This is just a test file to do some POC work on things like successfully splitting a file into multiple pieces and verifying that we can put it back together successfully (and also detect errors). I wouldn't be surprised if this code was eventually re-used in the actual tool, but it probably won't live in "chunktest.py".